damage: move the extension under the Xext directory
Tidying up the source tree structure a bit.
