+++ composite/compint.h
@@ -68,7 +68,7 @@
 #include "privates.h"
 #include "mi.h"
 #include "damage.h"
-#include "damageextint.h"
+#include "Xext/damage/damageextint.h"
 #include "xfixes.h"
 #include <X11/extensions/compositeproto.h>
 #include "compositeext.h"
